Add test for Huge MSBFS ASP as well

+++ algo/src/main/java/org/neo4j/graphalgo/impl/msbfs/HugeBfsConsumer.java
@@ -3,5 +3,5 @@ package org.neo4j.graphalgo.impl.msbfs;
 @FunctionalInterface
 public interface HugeBfsConsumer {
 
-    void accept(long nodeId, long depth, HugeBfsSources sourceNodeIds);
+    void accept(long nodeId, int depth, HugeBfsSources sourceNodeIds);
 }
